Update Listener with new read method
The Listener class should be portable to all types of message
structures, hence the read_until() method for serial ports is
used to specify the terminiation of a valid message.

In order to decouple the message structure specification and
listener functionality - ultimately to abstract away the more
complicated Listener class - the end_flag which is passed to the
read_until() method is specified in the Message classes.

Further, the Message classes should be a local defintion that any
user can define, with some core base structures defined for plug-
-n-play functionality.